Update files for JSON parsing
Updated the data access layer, dentists, and dentist registry to get it to parse the JSON file. This is still a work in progress as some things are still not working. The main thing is that the fields coordinates with longitude and latitude, and the fields openinghours with the weekdays need to be handled as separate objects, meaning that coordinates and openinghours need to be represented in the code as well.

Resolves #1