Publish changed dentists
Changed the USER_ID for both Publisher and Subscriber to 'dentist-' so that it will be easier to identify in mosquitto. Updated the README so that it is up to date with the repository. Added functionality for re-publishing dentists if there's an update to the dentists.json file. Using a watcher for this but if we cared about how it looks, we should be used the Observer pattern. Found examples of both ways and I included them in the comment above the method.

Resolves #5